backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18      # загрузка файлов

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0               # async PostgreSQL драйвер
alembic==1.14.0

# Cache / Queue
redis==5.2.1
celery==5.4.0
celery[redis]==5.4.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.6.1

# Storage (S3 / MinIO)
boto3==1.35.74
#aioboto3==13.2.0

# Document parsing
pymupdf==1.24.13              # PDF
python-docx==1.1.2            # DOCX
openpyxl==3.1.5               # XLSX
python-magic==0.4.27          # определение типа файла
Pillow==11.0.0                # изображения
pytesseract==0.3.13           # OCR

# LLM / AI
langchain==0.3.9
langchain-community==0.3.9
langchain-openai==0.2.11      # OpenAI-совместимый клиент (Ollama тоже)
openai==1.57.4                # базовый клиент
anthropic==0.40.0             # Anthropic Claude
chromadb==0.5.23              # векторная БД

# Output generation
python-docx==1.1.2
jinja2==3.1.4                 # шаблонизатор промптов

# Validation / Utils
pydantic==2.10.3
httpx==0.28.0
structlog==24.4.0             # структурированные логи
tenacity==9.0.0               # retry логика
python-dotenv==1.0.1
